day04: third attempt
Use a mutable array in the ST monad to try to simplify the construction
of the PunchedClocks.  This ends up coming to about the same amount of
code, but I feel guardClock is a little more readable than the functions
it replaces.

Who am I kidding, I just wanted to try out mutable arrays.