day04: second attempt
Use a map to store the time cards for all guards.  What was previously
called a time card is now called a punched clock.  The main functions
(sleepy, napTime, and sleepiestTime) now operate on the TimeCard
directly.  This avoids going over the event list a bunch of times, but
it doesn't seem to do much for run-time performance for the relatively
small inputs we have.