day10: solve puzzles
Use the sign of the derivative of the Y-range of the constellation to
detect when the message appears.  This assumes the message is visible
when the constellation is as vertically small as possible.

If we're being entirely honest, I first solved the problem by forwarding
roughly to where I knew the message would be assuming 8 units tall
letters (as in the example) and stepping forward.