day12: solve puzzles
The solution to the second half is based on the empirical fact that,
after a certain amount of time, the pattern just starts moving right pot
by pot but maintaining its shape, which produces an arithmetic series on
the sum of the numbers of the pots containing plants.  This was the case
for the AoC sample and my specific input, but after playing around a bit
with different rules this doesn't seem to be a necessary result for any
automaton with this type of rules.  I also looked for results on 1D
cellular automata which could help ground this, but I couldn't find
anything.